Use more "Modern" style skeleton in bob.pm
As per exercism/exercism.io#2043 , The skeleton is incredibly outdated.

The use of Exporter is not even necessary to complete the bob exercise,
as the exercise calls functions directly, not importing them.

But Exporter related code is left in to continue to serve as an example
of how to do it for people who may attempt to re-use the style.

Strictures/Warnings standards and version declaration standards, as well
as variable declaration standards are all updated to the state of the
art since March 2000.

Also, unlike the previous example, this change avoids any kind of "default" exporting,
as best practices state exporting should be done on request, not by default,
except in special cases.

https://metacpan.org/pod/Perl::Critic::Policy::Modules::ProhibitAutomaticExportation#DESCRIPTION

This patch should closes exercism/exercism.io#2043